+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.7.0] - 2018-12-21
+
 ### Added
 
 - Public test log event comparison function
@@ -92,7 +95,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/goph/logur/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/goph/logur/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/goph/logur/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/goph/logur/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/goph/logur/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/goph/logur/compare/v0.3.0...v0.4.0